fix(constant): move constant to Constants.js
